Repl: handle exceptions from within command handlers
very ugly; but, to be fair, Perl's string eval is also an exception
catcher; so, the eval part later on does catch them anyway already,
and we don't want to catch in two layers, do we?

Perl should perhaps provide a string eval that doesn't catch
exceptions (wrapper that re-throws them?).