PXML: interesting problem
Iff the element around the array (including the array, well, which
would have to be turned into a linked list first) could be let go,
then it would work. That means that all of the necessary properties
(element name) need to be get from the element first. Which should be
doable. (And make sure the holder of the whole tree lets go of it,
too, of course, but that's known.)