howto: should perhaps research how widespread it is / where it is spread
It even seems to have infected the Scheme community to some extent,
"we don't have classes (by default), [because|but] OO is not the
right/best idiom most of the time". Then Scheme has builtin type
hierarchies with generic ops but doesn't provide the means to build
the same as libraries out of the box.