List,Stream: make for_each work for mixed sequences
Do not introduce new method here since, it's the same, and I don't
want to add the alias everywhere.

So, this in the end probably *does* mean that I should drop
FP_Sequence_ref again, too. I just fear that unrelated methods with
the same name will be taken accidentally. (But maybe it's an accident
in an accident, so not that important? Or should non-sequence
improper_lists be safe?)