Linking: prefer CPAN url if available
Because CPAN actually does POD formatting. OK, there's many modules
without proper POD yet, todo.

Also, is this risky since it might now show things from CPAN that's
not ours? But that shouldn't happen, if it's found as a .pm file here,
it should have the namespace for getting its proper spot on CPAN thus
it will be ours. The exception is `meta/`. But there are only .pm
files below FunctionalPerl/ thus should be safe, too.