ModulePODUrl: remove predicate check again
Since there's some `3` somewhere in the docs apparently. Sure. That's
why I wanted to check for IV or NV. Not the string. Yes, can just
return () if it's not looking like a CPAN module name. But that is not
the thing that does the error reporting. Those are separate things. I
don't do the look-like-CPAN because CPAN will tell me anyway. But I
wanted the error report. Yes, overload fine, the ->'(""' shall be the
error checking then. But there's none for numbers.