Update docs on exception formatting
There is of course the temptation to partially undo this, show string
exceptions as strings; but, isn't it better to be unambiguous,
consistent, and always show the show'ed value? Can't tell if it's
really a string exception from just the string--OK, would need an evil
FP_Show_show function, though, right? Still, even then consistency
counts for something, or so.

Just because I can omit the quoting doesn't mean it's a good idea.(TM)